# BitVector::build_index: 100x speedup