Process the `BitVector` unit-by-unit instead of bit-by-bit.

Use `PopCount::count()` to update `num_1s` and use `select_bit` to find
the bit positions for the `select0s_` and `select1s_` indexes.

According to my benchmarks, the old bit-by-bit version processed a
256kbit vector at about 20MB/s independent of `enables_select0` and
`enables_select1`.

The new version is 50x-150x faster, depending on the compiler and build_index
options.

`enables_select_0=enables_select1=false`:
popcnt, no bmi2: 1600MB/s
popcnt, no bmi2: 2500MB/s
popcnt and bmi2: 2900MB/s

`enables_select_0=enables_select1=true`:
no popcnt, no bmi2: 1100MB/s
popcnt, no bmi2: 1600MB/s
popcnt and bmi2: 1800MB/s