@s-yata 2020-06-16
### Benchmark

- Data: enwiki-20191001-all-titles-in-ns0 (downloaded from https://dumps.wikimedia.org/enwiki/)
- Tool: marisa-benchmark

The following table shows build speed [1,000 keys/second].

#tries| s-yata:master [K/s]| jmr:build-index [K/s]
-----:|-------------------:|---------------------:
1     |            1,054.64|              1,087.85
2     |              915.17|                937.07
3     |              901.00|                920.59
4     |              896.08|                914.57
5     |              894.30|                912.47

jmr:build-index is 2-3% faster than s-yata:master.