@jmr 2020-06-16
> jmr:build-index is 2-3% faster than s-yata:master.

Did you configure with `--enable-native-code`? `popcnt` and `select_bit` are going to be important.

My benchmark was just on `BitVector::build_index`. I don't know what fraction of `marisa-benchmark` is spent in `build_index`, so I can't say whether more than 2-3% is expected.

I will have time to run/profile the benchmarks myself later in the week.