@s-yata 2020-06-16
The table shows the speed of dictionary construction and `BitVector::build_index` is not a major part of it.
However, I think the improvement is enough to accept this pull request.